Backend/requirements.txt
Django>=4.2.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.0
python-dotenv>=1.0.0
google-generativeai>=0.3.0
PyPDF2>=3.0.0
python-docx>=1.0.0
pydantic>=2.0.0 